Deliver basic no-frills signals
The most basic delivery of a signal to user. No fancy stuff.

+++ include/asm-i370/sigcontext.h
@@ -12,6 +12,6 @@ struct sigcontext_struct {
 	struct pt_regs 	*regs;
 };
 
-int i370_do_signal(sigset_t *oldset, struct pt_regs *regs);
+unsigned long i370_do_signal(sigset_t *oldset, struct pt_regs *regs);
 
 #endif
